// src/errors/app-error-code.enum.ts
export enum AppErrorCode {
  UNPROCESSABLE_ENTITY = 1,
  BAD_REQUEST = 3,
  GATEWAY_TIMEOUT = 4,
  NOT_FOUND = 5,
  CONFLICT = 6,
  FORBIDDEN = 7,
  TOO_MANY_REQUESTS = 8,
  PRECONDITION_FAILED = 9,
  REQUEST_TIMEOUT = 10,
  OUT_OF_RANGE = 11,
  NOT_IMPLEMENTED = 12,
  INTERNAL = 13,
  SERVICE_UNAVAILABLE = 14,
  UNAUTHORIZED = 16,
}